requirements.txt
google-auth==2.27.0
google-auth-oauthlib==1.2.0
gspread==5.12.4
oauthlib==3.2.0
simple-term-menu==1.6.4
colorama==0.4.6